Refactored the single event.go handlers file into separate files, makes code separation clearer
